move to node-testing/bench
