+++ bin/node/testing/src/lib.rs
@@ -21,4 +21,4 @@
 pub mod client;
 pub mod genesis;
 pub mod keyring;
-
+pub mod bench;
